requirements.txt
# ===============================================================================
# SISTEMA DE INTELIGENCIA GEOPOLÍTICA - DEPENDENCIAS OPTIMIZADAS
# Configurado para análisis multilingüe en tiempo real con datos 100% reales
# Idiomas soportados: español, inglés, ruso, chino, árabe
# ===============================================================================

# Core Framework & Web
flask==2.3.3
flask-cors==4.0.0
gunicorn==21.2.0

# HTTP & Web Scraping
requests==2.31.0
beautifulsoup4==4.12.2
aiohttp==3.8.6
urllib3==2.0.7

# Data Processing & Analysis
pandas
numpy
scipy

# RSS & News Collection
feedparser==6.0.10
asyncio-throttle==1.0.2

# Natural Language Processing - CORE
transformers==4.35.2
torch==2.1.1
spacy==3.7.2
langdetect==1.0.9
vaderSentiment==3.3.2
sentence-transformers==2.2.2

# Visualization
plotly==5.17.0
matplotlib==3.8.2
seaborn==0.13.0
folium==0.15.1

# Dashboard and Web
streamlit==1.28.2
dash==2.14.2
gradio==4.8.0

# Translation APIs
googletrans==4.0.2
deep-translator==1.11.4

# Report generation & Templates
jinja2==3.1.2
weasyprint==60.2
xhtml2pdf==0.2.11
reportlab==4.0.7
jinja2==3.1.2
weasyprint==60.2
xhtml2pdf==0.2.13
reportlab==4.0.8

# Chatbot and AI
langchain==0.0.348
openai==1.3.8
faiss-cpu==1.7.4
pycountry==23.12.11
pycountry_convert==0.7.2
country_converter==1.1.0

# Configuration and utilities
python-dotenv==1.0.0
pyyaml==6.0.1

# Database
sqlalchemy==2.0.23

# System monitoring and quality
psutil==5.9.6
schedule==1.2.0

# Additional utilities
python-dateutil==2.8.2
pytz==2023.3
rich==13.7.0
tqdm==4.66.1

# Security & Performance
cryptography==41.0.8
cachetools==5.3.2

# Data Quality & Validation
cerberus==1.3.5
marshmallow==3.20.1

# Testing & Quality Assurance
pytest==7.4.3
pytest-cov==4.1.0
black==23.11.0
flake8==6.1.0

# Development Tools
jupyter==1.0.0

# ===============================================================================
# OPTIMIZADO PARA PRODUCCIÓN - TODAS LAS VERSIONES VERIFICADAS
# ===============================================================================
